third_party/WebKit/LayoutTests/fast/forms/color/input-color-chooser-shown-readonly-expected.txt

PASS testRunner.isChooserShown() is true
PASS testRunner.isChooserShown() is false
PASS successfullyParsed is true

TEST COMPLETE
